Added basic implementation of daemon.
The daemon to pull data from Petfinder API can now pull the first
1000 dogs and insert them into a postgres database.  A transaction
is used to insert the first 1000 dogs.  A connection pool has also
been implemented.  Added api/sql folder that contains .sql files
that were used for the schema for this daemon.

Next steps are to abstract the call to the Petfinder API further
to pull > 1000 dogs in an automatic manner.