Added sql script to update specific dogs for Hugo.
Cleaning data from Petfinder API for three specific dogs as
requested by Hugo for the Demo Monday.  This is because Petfinder
API does not check whether dog names are valid and consequently
we get dog names that are named not real dog names.
Changing names of the below.
"DONATIONS: BEDS & TOYS" to "Bobby" for id 22406048.
"FOSTER HOMES" to "Lio" for id 32883666.
"TRANSPORT VOLUNTEERS!" to "Cujo" for id 31945123.