Renamed all Demo endpoints to Test.
Updated Demo endpoints to Test for testing purposes.  Test
endpoints are wrappers for the actual function calls that bypass
the requirement for access_token.  This is done by manually
setting the req.user.id and then calling the actual (non-test)
function.  The change is needed to show proper code coverage in
tests.