Added istanbul to package.json. Updated npm test.
Updated npm test command to call istanbul script in test mode.
To run the server type the command "npm start" into the shell.
To run the tests type the command "npm test" into the shell.
To run the tests with coverage type the command "npm test --coveage"
into the shell.